Factor out the LlvmAstParser into a separate file
