+++ lib/base-compiler.js
@@ -39,6 +39,7 @@ import { getDemanglerTypeByKey } from './demangler';
 import * as exec from './exec';
 import { InstructionSets } from './instructionsets';
 import { languages } from './languages';
+import { LlvmAstParser } from './llvm-ast';
 import { LlvmIrParser } from './llvm-ir';
 import { logger } from './logger';
 import { Packager } from './packager';
@@ -73,6 +74,7 @@ export class BaseCompiler {
 
         this.asm = new AsmParser(this.compilerProps);
         this.llvmIr = new LlvmIrParser(this.compilerProps);
+        this.llvmAst = new LlvmAstParser(this.compilerProps);
 
         this.toolchainPath = getToolchainPath(this.compiler.exe, this.compiler.options);
 
@@ -479,7 +481,7 @@ export class BaseCompiler {
         // A higher max output is needed for when the user includes headers
         execOptions.maxOutput = 1024 * 1024 * 1024;
 
-        return this.processAstOutput(
+        return this.llvmAst.processAst(
             await this.runCompiler(this.compiler.exe, newOptions, this.filename(inputFilename), execOptions));
     }
 
@@ -1079,87 +1081,6 @@ export class BaseCompiler {
         return compilerVersion.includes('clang') ||
             compilerVersion.match(/^([\w-]*-)?g((\+\+)|(cc)|(dc))/g) !== null;
 
-    }
-
-    processAstOutput(output) {
-        output = output.stdout;
-
-        // Top level decls start with |- or `-
-        const topLevelRegex = /^([`|])-/;
-
-        // Refers to the user's source file rather than a system header
-        const sourceRegex = /<source>/g;
-
-        // Refers to whatever the most recent file specified was
-        const lineRegex = /<line:/;
-
-        let mostRecentIsSource = false;
-
-        // Remove all AST nodes which aren't directly from the user's source code
-        for (let i = 0; i < output.length; ++i) {
-            if (output[i].text.match(topLevelRegex)) {
-                if (output[i].text.match(lineRegex) && mostRecentIsSource) {
-                    // do nothing
-                } else if (!output[i].text.match(sourceRegex)) {
-                    // This is a system header or implicit definition,
-                    // remove everything up to the next top level decl
-                    // Top level decls with invalid sloc as the file don't change the most recent file
-                    let slocRegex = /<<invalid sloc>>/;
-                    if (!output[i].text.match(slocRegex)) {
-                        mostRecentIsSource = false;
-                    }
-
-                    let spliceMax = i + 1;
-                    while (output[spliceMax] && !output[spliceMax].text.match(topLevelRegex)) {
-                        spliceMax++;
-                    }
-                    output.splice(i, spliceMax - i);
-                    --i;
-                } else {
-                    mostRecentIsSource = true;
-                }
-            }
-            // Filter out the symbol addresses
-            const addressRegex = /^([^A-Za-z]*[A-Za-z]+) 0x[\da-z]+/gm;
-            output[i].text = output[i].text.replace(addressRegex, '$1');
-
-            // Filter out <invalid sloc> and <<invalid sloc>>
-            let slocRegex = / ?<?<invalid sloc>>?/g;
-            output[i].text = output[i].text.replace(slocRegex, '');
-
-            // Unify file references
-            output[i].text = output[i].text.replace(sourceRegex, 'line');
-        }
-        const spanRegex = /<((?:line|col)[\d ,:ceilno]+)>/;
-        var lfrom, lto;
-        for (const element of output) {
-            let m = element.text.match(spanRegex);
-            if (m) {
-                let span = m[1];
-                const twoLineRegex = /line:(\d+)[,:].*line:(\d+)[:>]/;
-                let ll = span.match(twoLineRegex);
-                if (ll) {
-                    lfrom = Number(ll[1]);
-                    lto = Number(ll[2]);
-                } else {
-                    const leftLineRegex = /^line:(\d+)/;
-                    let left = span.match(leftLineRegex);
-                    if (left) {
-                        lfrom = Number(left[1]);
-                        lto = lfrom;
-                    } else {
-                        const rightLineRegex = / line:(\d+):/;
-                        let right = span.match(rightLineRegex);
-                        if (right) {
-                            lfrom = lto;
-                            lto = Number(right[1]);
-                        }
-                    }
-                }
-                element.source = { from : lfrom, to : lto };
-            }
-        }
-        return output;
     }
 
     async processGccDumpOutput(opts, result) {
